+++ NEWS.md
@@ -33,7 +33,7 @@
 * `inst/ChangeLog` is removed and replaced by `NEWS.md`.
 * `cmulti` models now have a `predict` method, see examples.
 * `"fmix"`" model type with time varying rates added to `cmulti` models.
-* Added new data set: `paired`.
+* Added new data sets: `paired`, `josm`, and `lhreg`.
 
 # Version 0.4-0, March 2, 2016