+++ NEWS.md
@@ -3,6 +3,7 @@
 * Use `Authors@R` field in `DESCRIPTION`.
 * Updated documentation.
 * Added SQPAD approach: see `?sqpad`.
+* Added new data sets: `paired`, `josm`, and `lhreg`.
 
 # Version 0.4-6, March 8, 2023
 
@@ -33,7 +34,6 @@
 * `inst/ChangeLog` is removed and replaced by `NEWS.md`.
 * `cmulti` models now have a `predict` method, see examples.
 * `"fmix"`" model type with time varying rates added to `cmulti` models.
-* Added new data sets: `paired`, `josm`, and `lhreg`.
 
 # Version 0.4-0, March 2, 2016